Moved QualifiedName into standalone class
