+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/Metrics.java
@@ -8,7 +8,7 @@ import static net.sourceforge.pmd.lang.java.oom.visitor.MetricsVisitorFacade.TOP
 
 import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
-import net.sourceforge.pmd.lang.java.ast.QualifiableNode.QualifiedName;
+import net.sourceforge.pmd.lang.java.ast.QualifiedName;
 import net.sourceforge.pmd.lang.java.oom.metrics.AtfdMetric;
 import net.sourceforge.pmd.lang.java.oom.metrics.WmcMetric;
